moved functions to a separate TU
